챗봇 메세지 커스텀 훅 리팩토링
